- Synchronize based also on id and uuid

+++ core/src/main/java/org/openmrs/sync/core/repository/OpenMrsRepository.java
@@ -5,12 +5,12 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.NoRepositoryBean;
 
 @NoRepositoryBean
-public interface OpenMrsRepository<E extends BaseEntity> extends JpaRepository<E, Integer> {
+public interface OpenMrsRepository<E extends BaseEntity> extends JpaRepository<E, Long> {
 
     /**
      * find entity by uuid
      * @param uuid the uuid
-     * @return entity
+     * @return an entity
      */
     E findByUuid(String uuid);
 }
